Initial commit for rippledmon:
`rippledmon` is a drop-in monitoring solution for
[`rippled`](https://github.com/ripple/rippled) instances.

`rippled is equipped to export metrics via `statsd` when
`insights` is configured. This monitoring tool consumes and
presents those metrics via statsd-graphite and Grafana docker
containers.

As more metrics are exported via `rippled`, corresponding
metrics collections and dashboards will be added to this
repository.